+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Change log
 - Added :meth:`text_scrubber.text_scrubber.TextScrubber.convert_html_entities`
 - Added :meth:`text_scrubber.text_scrubber.TextScrubber.fix_bad_unicode`
 - Added :meth:`text_scrubber.text_scrubber.TextScrubber.latex_to_text`
+- Added :meth:`text_scrubber.text_scrubber.TextScrubber.normalize_unicode`
 - Removed Python 3.6 support
 
 0.4.2